Move evilScenario to another file
